refactor: MapLetterService 클래스 역할별 분리
